+++ src/constants.ts
@@ -6,7 +6,7 @@ export const SIMULCAST = ["", "true", "false"] as const;
 
 export const SPOTLIGHT = ["", "true", "false"] as const;
 
-export const AUDIO_CODEC_TYPES = ["", "OPUS"] as const;
+export const AUDIO_CODEC_TYPES = ["", "OPUS", "LYRA"] as const;
 
 export const AUDIO_BIT_RATES = ["", "8", "16", "24", "32", "64", "96", "128", "256"] as const;
 
@@ -48,7 +48,7 @@ export const RESOLUTIONS = [
 
 export const DISPLAY_RESOLUTIONS = ["", "480p (640x480)", "240p (320x240)"] as const;
 
-export const FRAME_RATES = ["", "60", "30", "24", "20", "15", "10"] as const;
+export const FRAME_RATES = ["", "60", "30", "24", "20", "15", "10", "5"] as const;
 
 export const AUTO_GAIN_CONTROLS = ["", "true", "false"] as const;
 
@@ -87,6 +87,8 @@ export const CONNECTION_STATUS = ["initializing", "disconnected", "disconnecting
 export const ROLES = ["sendrecv", "sendonly", "recvonly"] as const;
 
 export const FACING_MODES = ["", "front", "back"] as const;
+
+export const LYRA_PARAMS_BITRATES = ["", "3200", "6000", "9200"] as const;
 
 export const WORKER_SCRIPT = `
 self.onmessage = (event) => {